third_party/blink/web_tests/external/wpt/cobalt/cobalt-system/cobalt-system.https.sub.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test CobaltSystem web API.</title>
  <script src='/resources/testharness.js'></script>
  <script src='/resources/testharnessreport.js'></script>
</head>
<body>
  <script>
    test(() => {
        assert_equals(window.h5vcc.cobaltSystem.advertisingId, "fake advertisingId");
    }, "cobaltSystem has the right property value");
  </script>
</body>
</html>
